cy without dot in the middle of disc

+++ includes/draw.h
@@ -6,7 +6,7 @@
 /*   By: rdanyell <rdanyell@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 12:06:29 by marlean           #+#    #+#             */
-/*   Updated: 2022/07/06 17:18:00 by rdanyell         ###   ########.fr       */
+/*   Updated: 2022/07/12 11:49:07 by rdanyell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,9 +40,9 @@ float vector_sumpow2(t_coord a);
 float vector_scalar(t_coord a, t_coord b);
 t_coord  vec3_mul(t_coord a, t_coord b);
 
-int sphere_intersect(t_camera cam, t_coord ray, t_sphere *sphere);
-int	plane_intersect(struct s_camera cam, t_coord ray,  t_plane *plane);
-int	cylindr_intersect(struct s_camera cam, t_coord ray,  t_cylind *cylind);
+float sphere_intersect(t_camera cam, t_coord ray, t_sphere *sphere);
+float	plane_intersect(struct s_camera cam, t_coord ray,  t_plane *plane);
+float cylindr_intersect(struct s_camera cam, t_coord ray,  t_cylind *cylind);
 
 int set_color(t_color color, float l);
 #endif
